requirements.txt
loguru==0.6.0
pytest==6.2.5
PyYAML==6.0
ruamel.base==1.0.0
selenium==3.8.1
